+++ lib/src/utils/name_transform_utils.dart
@@ -1,5 +1,0 @@
-String camelToCapitalized(String name) {
-  return name.replaceAllMapped(RegExp(r'^([a-z])|[A-Z]'),
-    (Match m) => m[1] == null ? ' ${m[0]}' : m[1]!.toUpperCase()
-  );
-}